dattx-so-productpage fix
dattx-so-productpage fix
